pyproject.toml
[tool.poetry]
name = "flashboard-tts-proto"
version = "0.1.0"
description = ""
authors = ["Federico Maggi <federico.maggi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.20.45"
pygame = "^2.1.2"
paho-mqtt = "^1.6.1"
cachetools = "^5.0.0"
dynaconf = "^3.1.7"
asyncio-mqtt = "^0.12.1"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
ipython = "^8.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
